+++ package.json
@@ -14,6 +14,7 @@
     "@react-router/serve": "^7.7.1",
     "github-slugger": "^2.0.0",
     "isbot": "^5.1.27",
+    "lucide-react": "^0.542.0",
     "react": "^19.1.0",
     "react-dom": "^19.1.0",
     "react-markdown": "^10.1.0",
